// src/types/dto/Wallet.ts
export interface WalletResponse {
	id: string;
	customer_id: string;
	name: string;
	currency: string;
	description: string;
	balance: string;
	credit_balance: string;
	wallet_status: WALLET_STATUS;
	metadata: Metadata;
	auto_topup?: {
		enabled: boolean;
		threshold: string;
		amount: string;
		invoicing: boolean;
	};
	wallet_type: WALLET_TYPE;
	config: {
		allowed_price_types: WALLET_CONFIG_PRICE_TYPE[];
	};
	conversion_rate: string;
	topup_conversion_rate?: string;
	created_at: string;
	updated_at: string;
	alert_settings?: WalletAlertSettings;
	alert_state?: WalletAlertState;
}